feat: Convert frontend to NextJS
