--- frontend/postcss.config.js
@@ -1,1 +1,0 @@
-module.exports = { plugins: { tailwindcss: {}, autoprefixer: {} } };
