--- frontend/tailwind.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  content: ["./index.html", "./src/**/*.{js,jsx}"],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-};
